Explicitly cast to uint64_t when creating a 64-bit data from 16-bit data.
Shifting uint16_t to the left without casting will be promoted once to
int32_t and then to uint64_t. When MSB of uint16_t us 1, int32_t
becomes negative, resuling in all the upper 32bit of uint64_t being 1.

+++ f3rp61/src/devAiF3RP61.c
@@ -315,10 +315,11 @@ static long read_ai(aiRecord *precord)
     if (option == 'D') {
         double val;
         if (device == 'A') {
-            uint64_t lval = (((int64_t)ldata[1])<<32) | ldata[0];
+            uint64_t lval = (((uint64_t)ldata[1])<<32) | (uint64_t)ldata[0];
             memcpy(&val, &lval, sizeof(double));
         } else {
-            uint64_t lval = (((uint64_t)wdata[3])<<48) | (((uint64_t)wdata[2])<<32) | (wdata[1]<<16) | wdata[0];
+            uint64_t lval = (((uint64_t)wdata[3])<<48) | (((uint64_t)wdata[2])<<32) | ((uint64_t)wdata[1]<<16) | (uint64_t)wdata[0];
+
             memcpy(&val, &lval, sizeof(double));
         }
         // todo : consider ASLO and AOFF field
